interface for debugging, Monster class
Added an interface that Enemy classes can use to print debug information which can be displayed under enemies later. For now, A separate bool can be toggled using the tilde button to allow the user to see the game's current state. Also added an early framework for the Monster class.