Slight change to collision system, Ghost
This push adds an additional Rectangle to entities for collision detection purposes. In practice, this makes it so their upper bodies can now be displayed over tables and the wall when they walk from under it, instead of having their head bump the bottom of them. Also included is a change to the monster pathfinding system so they are slightly better at pathing around obstacles, and changes to the Ghost class so that they are invincible for longer and can move after being hit(as well as a visual indicator to differentiate between them) and attack more slowly to make it fairer to the player.